refactor(frontend): move Products component to elements folder and update imports
