+++ packages/frontend/src/components/containers/index.ts
@@ -1,9 +1,8 @@
 import RegisterForm from './RegisterForm';
 import LoginForm from './LoginForm';
-import Products from './Products';
 import PayPal from './PayPal';
 import CartContent from './CartContent';
 import UserForms from './UserForms';
 import UserProducts from './UserProducts';
 
-export { RegisterForm, LoginForm, Products, PayPal, CartContent, UserForms, UserProducts };
+export { RegisterForm, LoginForm, PayPal, CartContent, UserForms, UserProducts };
