+++ packages/frontend/src/components/elements/index.ts
@@ -3,5 +3,6 @@ import Input from './Input';
 import Loader from './Loader';
 import PrimaryButton from './PrimaryButton';
 import Product from './Product';
+import Products from './Products';
 
-export { Nav, Input, Loader, PrimaryButton, Product };
+export { Nav, Input, Loader, PrimaryButton, Product, Products };
