+++ packages/frontend/src/components/pages/Home.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 
 // import { Searcher } from '../molecules';
-import { Products } from '../containers';
+import { Products } from '../elements';
 import { Nav } from '../elements/';
 import { useAuth } from '../../hooks';
 import Cookies from 'universal-cookie';
